Move ChecksCatalog component to pages directory
