+++ assets/js/pages/ChecksCatalog/index.js
